Read extended desc
Built up more Oauth methods. (The demo script is very under developed
requiring you to fill in many of the methods enforced by the interfaces)
Created a series of example routes to make development debuging a bit
easier. (Not to replace PHPUnit, but to debug little things as I go
along)